extracting StackTrace utility class
